+++ src/main/java/com/mrbysco/densetrees/block/DenseLogBlock.java
@@ -1,5 +1,6 @@
 package com.mrbysco.densetrees.block;
 
+import com.mrbysco.densetrees.registry.DenseRegistry;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
 import net.minecraft.core.particles.ParticleTypes;
@@ -34,8 +35,9 @@ public class DenseLogBlock extends RotatedPillarBlock {
 		super.animateTick(state, level, pos, random);
 
 		if (level.isClientSide) {
+			double yOffset = state.is(DenseRegistry.DENSE_MANGROVE_LOG.get()) ? -1.5D : 0;
 			level.addParticle(ParticleTypes.GLOW,
-					(double) pos.getX() + 0.5D, (double) pos.getY() + 0.5D, (double) pos.getZ() + 0.5D,
+					(double) pos.getX() + 0.5D, (double) pos.getY() + 0.5D + yOffset, (double) pos.getZ() + 0.5D,
 					(double) ((float) (level.random.nextFloat() - 0.5) * 2 + random.nextFloat()) - 0.5D,
 					(double) ((float) (level.random.nextFloat() - 0.5) * 2 - random.nextFloat() - 1.0F),
 					(double) ((float) (level.random.nextFloat() - 0.5) * 2 + random.nextFloat()) - 0.5D);