+++ changelog.md
@@ -1,1 +1,2 @@
-* Initial release
+* Initial 1.19.2 release
+* Added Dense Mangrove trees